Initial commit
Status: Still in progress
Current Issues: I have set up the calculateNearMiss class, currently takes in the integer value for power and the integer value for the upper limit. We still need to figure out how to properly compare and find the near miss difference. When calculating the difference, sometimes I get a negative value, which is to be expected based on how I set this up, so we need to figure out a way to avoid having this issue.